2979054: Support for 0..n plural forms instead of only 1.
The number of plural forms is hard-coded in the filter for all languages.
(If a language is missing, the default is 1 plural form)
If for a given language the number is n, then the translated po-file contains
msgstr[0] ".." till msgstr[n] ".." lines. The number of msgstr[x] lines
in the source file does not matter. if the source contains more lines than n,
then the extra lines are ignored (and not aligned).

git-svn-id: https://omegat.svn.sourceforge.net/svnroot/omegat/trunk@4635 b0d8beef-cb45-0410-a8e4-c0d495c3b779