Tokenizer token cache should not be static
When two languages happen to have words with the same orthography
(e.g. "promotion" in English and French), if the caches are shared
across instances (are static) then the cached results for one language
will be returned for the other language. This is incorrect, though
usually only noticeable when stemming.

git-svn-id: svn+ssh://svn.code.sf.net/p/omegat/svn/trunk@10184 b0d8beef-cb45-0410-a8e4-c0d495c3b779