Mark paragraphs delimitation
Added an option to mark the beginning of paragraphs, with the possibility to customize the text and color of this delimitation. Even when this option is turned off in the menu, a paragraph mark is added in the numbering at the end of the active segment.

This feature is ported from Thomas Cordonnier's OmegaT-DGT 3.2.

RFE#431 and RFE#314

git-svn-id: svn+ssh://svn.code.sf.net/p/omegat/svn/trunk@10539 b0d8beef-cb45-0410-a8e4-c0d495c3b779