bug#860 Queue scripts executed during events
When executing scripts during events, the scripts are sorted by
filename, the corresponding SwingWorkers are created and queued.

The elements of the queue are then executed and removed from the queue
when done. The state of the SwingWorker is monitored with the
PropertyChangeListener to check the state property is DONE.

As a side-effect, this fix also removes the unwanted canceled (0ms) that
were appearing when several scripts were launched in rapid succession.