Fix Win installer fallback to wrong language for custom messages
E.g. zh-TW doesn't have custom messages; apparently Inno Setup falls back to the
first-defined language, which happened to be Russian in my test.

The order of the message definitions depends on the iteration order of the
fileTree, which is probably undefined. Sorting the tree gets English to the top
because it has an empty language code by our convention.