Changing the flow of updating teamwork omegat.project
This patch changes the flow of updating local teamwork
"omegat.project" file with the content of remote copy of
"omegat.project" on the git server.

Before this patch, the content of remote copy overwrites the local
copy first, and then if the content lacks the git setting, it was
added and then the "omegat.project" was updated to reflect the
addition.

However, if OmegaT terminated early, such update may not happen and
thus rendering the teamwork without proper git setting.  When the
remote copy of "omegat.project" lacks git setting, on the next
opening, the project may no longer be a teamwork project (!).  This
confuses the end users to no end.

This patch changes the workflow of update as follows.

1. Content of the remote copy of "omegat.project" on the server is
   written to local file, "omegat.project.NEW".

2. The content of this file is parsed.

3. If git-setting is missing in the content, it is added from then
   current setting from the local copy of "omegat.project".

4. Only then we create a backup copy of the old "omegat.project" in
   omegat.project.TIMESTAMP.bak.  Then we RENAME "omegat.project.NEW"
   to "omegat.project".

With this flow, there is always a working copy of omegat.project on
the local computer even if OmegaT terminates early.  So we don't end
up with incorrect non-teamwork copy even if the remote copy of
"omegat.project" lacks git setting.

Also, this patch displays a warning dialog to warn the user when the
remote copy of "omegat.project" on the git server lacks git setting.
This should help end users to detect the misconfiguration early on.

modified:   src/org/omegat/core/team2/RemoteRepositoryProvider.java
        Update two helper functions that can copy the
        remote "omegat.project" to local "omegat.project.NEW".
        These need to emulcate the functions of existing functions
        to find the location of remote copy of "omegat.project"
        correctly in the presence of "mapping".
modified:   src/org/omegat/gui/main/ProjectUICommands.java
        Change of workflow as described above, and add a couple of
        log lines.

- Add test case to copy omegat.project to omegat.project.NEW

- Team: remove dead code
	Remove a code chunk that is commented out in v4.0 in Mar. 2016 by alex73
	at svn+ssh://svn.code.sf.net/p/omegat/svn/trunk@8304
	which is f71b80ab13cbb34847258f26722fac138e6ca94a in git.

- Team(git): always respect a specified team URL

  When team project has a mapping configuration, there is a default repository URL in omegat.project.
  There are two URL protocol schema, git+ssh or https, for same repository.
  This patch check whether a repository specified is git type then, always override
  repository URL with user specified to clone.
  For example, when team project repository is configured "https://github.com/OmegaT-L10N/ja"
  but user specify "git@github.com:OmegaT-L10N/ja", get omegat.project using ssh+git and
  and override a value "https://github.com/OmegaT-L10N/ja" by "git@github.com:OmegaT-L10N/ja".
  Then OmegaT continues to use git+ssh for the repository.

Signed-off-by: Hiroshi Miura <miurahr@linux.com>