Link and indent fixes
I restored the indent to match the one in the current master, and changed the wording of the section on event-based scripts a little bit.
Adding the link again seemed redundant when we already have to links to that section of the manual immediately above, so I removed it entirely in favour of an indirect reference.

Question: shouldn't we also mention the `properties` subfolder for scrips that are localized?
(Too much crosslinking becomes tedious for the reader/user, especially when the same not-so-informative sub-section is linked three times within four sentences...)