fix(gui/filter): wrong target selection on filter customizer table
We add a sorter at commit bb8f181998e73a1b023e9e3942d84830e18cacd8
When the table is sorted, the indices of the rows in the view don't
match with the indices in the model anymore.
This add `convertRowIndexToModel` method to convert from a sorter index to
model index.

ref: https://sourceforge.net/p/omegat/mailman/message/37770101/
Signed-off-by: Hiroshi Miura <miurahr@linux.com>