chore: tweak jcache dependency
Caffeine 3.1.0 and before use shadowJar plugin that produce both shadowRuntimeElement and runtimeElement.
It is a cause of dependency resolution problem for projects that use OmegaT as a library.
Caffeine 3.x requires Java 11, so we cannot bump it to 3.1.1 and later.

We have already treated it for caffeine library, but the problem is also in jcache library.

Signed-off-by: Hiroshi Miura <miurahr@linux.com>