feat(newUI): use different label for quick access tool on macOS
New UI component consist with burgerMenu and quick access tool is
available on Windows and Linux. A macOS has limitation on menu system,
so it cannot support a smart menu.
We put quick access tool as a tool bar on macOS.
Two feature is similar but different. It is because we use different
label for the feature among Linux/Windows and macOS.

Signed-off-by: Hiroshi Miura <miurahr@linux.com>