[BUGS#1164] fix: break project translation when rebase and commit
When saving a team project and another member have updated translations
and OmegaT run rebaseAndCommit, enforce TM become removed from
the result.

The change here to change that when merge TMX from remote and local, we
save it into file, but DON'T replace it in projectTMX.
We reload the file when necessary, in RebaseAndCommit.IRebase#reload.

It changes RealProject#mergeTMX to return merged TMX
and the caller should handle the tmx.

Signed-off-by: Hiroshi Miura <miurahr@linux.com>