docs(devel): add guideline for welcoming beginners note
There are many issue tickets and bug reports in source forge issue manager
and dev mailing list from beginners or not well-exprienced users.

Many of it does not have enough information to reproduce the issue and identify a target
source code. To reply and support these post consumes limited times and motivations of
contributors.
To help these members, I'd like to add a guidelines and text snippet to reply easily.

Signed-off-by: Hiroshi Miura <miurahr@linux.com>