chore: update linux-deb/rpm build
A jpackage plugin author declined to fix the bug on gradle cache feature with my proposal to change, and express the project won't support cache feature.
The plugin has no more benefit other than a syntax sugar, so I decided to remove the plugin and migrate it as Exec task type.

Now we can enable a configuration cache and speed-up our build.

Signed-off-by: Hiroshi Miura <miurahr@linux.com>