fix: docker/nerdctl detection
When using Rancher-desktop for container execution, it installs both docker and nerdctl commands.
Because a execution of docker command with Rancher Desktop fails, check a status of `docker info` command, and wen failed, we select nerdctl.
nerdctl can also be failed when user forget to start Rancher desktop.
We run ``$CMD info` after the selection, then user can recognize a error with warning message to launch Rancher desktop.

The change does not affect other environment, beucase we use it when docker info or nerdctl info succeeded.

Signed-off-by: Hiroshi Miura <miurahr@linux.com>