fix: restartGUI when jpackage distribution
jpackage distribution don't have `bin/java` executable.
It need to find a launcher `bin/OmegaT` and execute it.
This change is to check existence of `$JAVA_HOME/bin/java`,
and when not found, try to check
 `<OmegaT installed folder/bin/OmegaT`.
If both java or launcher exectable are not found, the attempt is aborted and show waring message.

Signed-off-by: Hiroshi Miura <miurahr@linux.com>