[BUGS#858] change alternate modifier to alt
Java Swing menu key accelerator is set with `Ctrl-<Key>` or `Ctrl-Alt-<Key>`(PC) or `Meta-<Key>` or `Meta-Alt-<Key>` (Mac), so accessing menu item with shortcut key always with a `Ctrl`/`Meta` modifier which is defined as `MenuShortcutKeyMask`.

OmegaT menu handler checks  whether a modifier is  `MenuShortcutKeyMask`.
It is always "true" ,  when accessed from a "shortcut key"

A change is intended to behave
1. Use the menu item normally to open the file
2. Hold "Alt" while using the menu item to reveal the file
3. Shortcut key "Ctrl+<key>", "Ctrl+<mod>+<key>" open the file

Signed-off-by: Hiroshi Miura <miurahr@linux.com>