[BUGS#1255] Fix: loading stardict3 w/ synonyms file
This is a backport from master branch.

stardict-JMDict is a Japanese-English dictionary in stardict format version 3.0.0 with synonyms.
It is a bug in stardict4j library when dictionary has synonyms file.
stardict4j project relases v0.6.3 with the fix.

It also shows ugly CSS contents in dictionary pane.
The patch add HTML filter.

Signed-off-by: Hiroshi Miura <miurahr@linux.com>