feat: filter(po): plural as comment in japanese target (#1165)
* feat/filter(po): plural as comment in japanese

Japanese is nplural=1 language.
When there is no marker in singular msgid, translator can not see plural form of msgid when translate in OmegaT panes.
This add plural form of msgid as a comment of a segment which is a singular form.

Signed-off-by: Hiroshi Miura <miurahr@linux.com>

* fix: test: Update expectations

* Fix typo in date

Might as well fix this typo too while we're at it.

---------

Signed-off-by: Hiroshi Miura <miurahr@linux.com>
Co-authored-by: kazephil <kazephil@gmail.com>