Add unit and functional test for o.o.g.e.DocumentFilter3 class (#1348)
* Refactor access to `Document3` properties with getter/setter methods.

- Replaced direct field access in `Document3` with private fields and corresponding getter/setter methods to improve encapsulation.
- Adjusted references across classes to utilize these methods, ensuring better code maintainability and adherence to object-oriented principles.

Signed-off-by: Hiroshi Miura <miurahr@linux.com>

* Add unit tests for DocumentFilter3 replace method

Introduce comprehensive tests for various scenarios of the replace method in DocumentFilter3, including valid replacements, out-of-bounds conditions, and trusted mode operations. This ensures robust validation of text editing logic within the editor.

Signed-off-by: Hiroshi Miura <miurahr@linux.com>

* Add Mockito library to test dependencies

Mockito has been added to the test dependencies to facilitate mocking in unit tests. The required version has also been defined in the `libs.versions.toml` file to ensure consistency across the project.

Signed-off-by: Hiroshi Miura <miurahr@linux.com>

* style: Remove unused import in DocumentFilter3Test

The `eq` static import was removed as it is not used in the test class. This change helps to clean up unused dependencies, maintaining code clarity and reducing potential confusion.

Signed-off-by: Hiroshi Miura <miurahr@linux.com>

---------

Signed-off-by: Hiroshi Miura <miurahr@linux.com>