Describe tmx prop display syntax in the manual.
Since only accidentally found based on code that this exists, and can be
useful.

Could imagine some follow-up improvements, like including an example
template in the dropdown of templates (so people skipping the manual
will encounter the syntax too), or adding template syntax enchancement
of being able to rename the prop type on display etc.